+++ src/test/java/calculator/InputProcessorTest.java
@@ -43,28 +43,28 @@ public class InputProcessorTest {
     @Test
     void 숫자_배열_출력_기본_구분자() {
         assertSimpleTest(() -> {
-            assertThat(processor.processInput("1,2:3")).isEqualTo(new int[]{1, 2, 3});
+            assertThat(processor.processInput("1,2:3")).isEqualTo(new long[]{1, 2, 3});
         });
     }
 
     @Test
     void 숫자_배열_출력_커스텀_구분자() {
         assertSimpleTest(() -> {
-            assertThat(processor.processInput("//*\\n1*2*3")).isEqualTo(new int[]{1, 2, 3});
+            assertThat(processor.processInput("//*\\n1*2*3")).isEqualTo(new long[]{1, 2, 3});
         });
     }
 
     @Test
     void 숫자_배열_출력_커스텀_구분자_2() {
         assertSimpleTest(() -> {
-            assertThat(processor.processInput("//*$\\n1*$2*$3")).isEqualTo(new int[]{1, 2, 3});
+            assertThat(processor.processInput("//*$\\n1*$2*$3")).isEqualTo(new long[]{1, 2, 3});
         });
     }
 
     @Test
     void 숫자_배열_출력_커스텀_구분자_빈문자() {
         assertSimpleTest(() -> {
-            assertThat(processor.processInput("//&\\n1&&2&3")).isEqualTo(new int[]{1, 0, 2, 3});
+            assertThat(processor.processInput("//&\\n1&&2&3")).isEqualTo(new long[]{1, 0, 2, 3});
         });
     }
 
@@ -87,7 +87,7 @@ public class InputProcessorTest {
     @Test
     void 숫자_배열_출력_빈_커스텀_구분자() {
         assertSimpleTest(() -> {
-            assertThat(processor.processInput("//\\n123")).isEqualTo(new int[]{1, 2, 3});
+            assertThat(processor.processInput("//\\n123")).isEqualTo(new long[]{1, 2, 3});
         });
     }
 }
